requirements.txt
argh==0.26.2
cassandra-driver==3.24.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dramatiq==1.9.0
Flask==1.1.2
geomet==0.2.1.post1
gevent==20.6.2
greenlet==0.4.16
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pathtools==0.1.2
pika==1.1.0
prometheus-client==0.8.0
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
urllib3==1.25.9
watchdog==0.8.3
watchdog-gevent==0.1.0
Werkzeug==1.0.1
zope.event==4.4
zope.interface==5.1.0
injector~=0.18.3